Added recycler in the daily weather fragment.

+++ app/src/main/java/com/example/myweatherapp/model/Coord.kt
@@ -4,6 +4,6 @@ package com.example.myweatherapp.model
 import com.google.gson.annotations.SerializedName
 
 data class Coord(
-    val lat: Double,
-    val lon: Double
+    val lat: Float,
+    val lon: Float
 )
