+++ app/src/main/java/com/example/myweatherapp/model1/Temp.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Temp(
     val day: Double,
     val eve: Double,
-    val max: Double,
+    val max: Float,
     val min: Double,
     val morn: Double,
     val night: Double
